requirements.txt
crewai
langgraph
fastapi
uvicorn
python-dotenv
pydantic
requests
langdetect